Support for MB/GB daily difference
This patch takes into consideration the possible differences for smaller caching servers. One day it may serve in MB and one day it may serve in GB or vice verse. 

This patch may give wrong results if one set of bandwidth is in GB and one is in MB on the same bandwidth line. I may or may not put more effort into this as this is primarily designed for larger cache servers that need this data.