Cacher 2.0
Simplifies code by removing repetitive echo (echo to screen + echo to file); replaces repetitive echo with tee as suggested by @hunty1 on PR 18 ( 2>&1 | tee -a "${tmplocation}"/AlertInfo.txt).  Previously, Cacher was 882 lines of code, this change brings it down to 744 and is easier to read and modify.

Adds Uptime to emailed report.

Adjusts Last Modified date from 05/16/2016 (inaccurate) to 9/23/2016.

Adds a script version, randomly started at 2.0. I assume that's a good point to start incrementing the script version for all future PR's and updates.
Adds script version to the bottom of the emailed alert for reference when comparing reports.